refactor: remove leftover debug and dead helpers
